backend/requirements.txt
# NPC Dialogue Web - Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
numpy>=1.24.0
openai-whisper>=20231117
python-dotenv>=1.0.0
